* commit '2b48459ead72dbb80073b6705722636a73f85eac': (36 commits)
  MALI: valhall: upgrade DDK to g29p1-12eac0, from g29p1-11eac0
  drm/rockchip: vop2: fix msmart grid overlap check error
  drm/rockchip: vop2: alloc correct size memory for grid scan point
  video: rockchip: vehicle: csi dphy add continue mode config
  UPSTREAM: ext4: clear i_state_flags when alloc inode
  soc: rockchip: thunderboot_sfc: Clear DMA int status
  media: platform: rockchip: vicap add yuv422-10bit support
  phy: rockchip: usbdp: Fix null dev pointer on DT tune sequence get
  input: touchscreen: hxchipset: Compiling himax_mmi.ko does not require enabling CONFIG_KALLSYMS_ALL
  arm64: configs: rk3576/rk3588 vehicle: enable compilation of himax_mmi.ko
  arm64: dts: rockchip: Fix touch and display anomalies for rk3576/rk3588 vehicle evb
  mtd: spinand: Remove ebh command support
  mtd: spinand: Kingston: Support new device SPI004-SDEG
  ASoC: rockchip: sai: Add property for FPC mode
  ASoC: rockchip: sai: Fix tdm slot mask when property absent
  drm: rockchip: dw_hdmi: Fix incorrect error that occurs when switching VRR type
  drm/rockchip: dw_hdmi: Fix gaming-vrr vfp can't be configured
  drm/rockchip: vop2: remove "inline" from vop2_mask_write()
  media: rockchip: vicap: virtual width align 24 byte for rgb888 to support more resolutions
  drm/rockchip: vop2: fix null point when use cursor layer
  ...

+++ drivers/phy/rockchip/phy-rockchip-inno-hdmi-phy.c
@@ -1573,7 +1573,7 @@ inno_hdmi_rk3528_phy_pll_recalc_rate(struct inno_hdmi_phy *inno,
 		frac = inno_read(inno, 0xd3) |
 		       (inno_read(inno, 0xd2) << 8) |
 		       (inno_read(inno, 0xd1) << 16);
-		vco += DIV_ROUND_CLOSEST(parent_rate * frac, (1 << 24));
+		vco += DIV_ROUND_CLOSEST((u64)parent_rate * frac, (1 << 24));
 	}
 	if (inno_read(inno, 0xa0) & 2) {
 		do_div(vco, nd * 5);
